requirements.txt
matplotlib==3.7.0
numba==0.56.4
numpy==1.21.5
Pillow==9.5.0
